# Conflicts:
#	src/main/java/com/example/CorporateLearningManagmentSystem/dto/CourseDetails.java
#	src/main/java/com/example/CorporateLearningManagmentSystem/dto/ModuleDetails.java
#	src/main/java/com/example/CorporateLearningManagmentSystem/entity/Course.java
#	src/main/java/com/example/CorporateLearningManagmentSystem/entity/Module.java
#	src/main/java/com/example/CorporateLearningManagmentSystem/entity/Resource.java

+++ src/main/java/com/example/CorporateLearningManagmentSystem/dto/CourseDetails.java
@@ -7,14 +7,14 @@ import java.util.List;
 
 
 public class CourseDetails {
-    private Long courseId;
+    private int courseId;
     private String name;
     private String description;
     private int duration;
 
     private List<ModuleDetails> moduleDetails;
 
-    public CourseDetails(Long courseId, String name, String description, int duration) {
+    public CourseDetails(int courseId, String name, String description, int duration) {
         this.courseId = courseId;
         this.name = name;
         this.description = description;
@@ -29,11 +29,12 @@ public class CourseDetails {
         this.moduleDetails = moduleDetails;
     }
 
-    public Long getCourseId() {
+
+    public int getCourseId() {
         return courseId;
     }
 
-    public void setCourseId(Long courseId) {
+    public void setCourseId(int courseId) {
         this.courseId = courseId;
     }
 
@@ -61,7 +62,7 @@ public class CourseDetails {
         this.duration = duration;
     }
 
-    public CourseDetails(Long courseId, String name, String description, int duration, List<ModuleDetails> moduleDetails) {
+    public CourseDetails(int courseId, String name, String description, int duration,List<ModuleDetails> moduleDetails ) {
         this.courseId = courseId;
         this.name = name;
         this.description = description;
