+++ src/main/java/com/example/CorporateLearningManagmentSystem/dto/ModuleDetails.java
@@ -7,9 +7,10 @@ import java.util.List;
 
 
 public class ModuleDetails {
-    private Long moduleId;
-//    private Course course;
+    private int moduleId;
+    //private Course course;
     private String moduleName;
+
 
     private List<ResourceDetails> resourceDetails;
 
@@ -24,11 +25,11 @@ public class ModuleDetails {
     public ModuleDetails() {
     }
 
-    public Long getModuleId() {
+    public int getModuleId() {
         return moduleId;
     }
 
-    public void setModuleId(Long moduleId) {
+    public void setModuleId(int moduleId) {
         this.moduleId = moduleId;
     }
 
@@ -48,14 +49,16 @@ public class ModuleDetails {
         this.moduleName = moduleName;
     }
 
-    public ModuleDetails(Long moduleId, String moduleName, List<ResourceDetails> resourceDetails) {
+    public ModuleDetails(int moduleId, String moduleName, List<ResourceDetails> resourceDetails) {
         this.moduleId = moduleId;
         this.moduleName = moduleName;
         this.resourceDetails = resourceDetails;
     }
 
-    public ModuleDetails(Long moduleId, String moduleName) {
+
+    public ModuleDetails(int moduleId, String moduleName) {
         this.moduleId = moduleId;
+//        this.course = course;
         this.moduleName = moduleName;
     }
 }
