+++ src/main/java/com/example/CorporateLearningManagmentSystem/entity/Module.java
@@ -10,7 +10,7 @@ public class Module {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long moduleId;
+    private int moduleId;
 
     @ManyToOne
     @JoinColumn(name = "course_id")
@@ -21,6 +21,10 @@ public class Module {
 
     @OneToMany(mappedBy = "module")
     private List<CourseModuleResourceMapping> mappings;
+
+
+
+
 
     @OneToMany(mappedBy = "module", cascade = CascadeType.ALL)
     private List<Resource> resources;
@@ -33,11 +37,11 @@ public class Module {
         this.resources = resources;
     }
 
-    public Long getModuleId() {
+    public int getModuleId() {
         return moduleId;
     }
 
-    public void setModuleId(Long moduleId) {
+    public void setModuleId(int moduleId) {
         this.moduleId = moduleId;
     }
 
@@ -65,7 +69,7 @@ public class Module {
         this.mappings = mappings;
     }
 
-    public Module(Long moduleId, Course course, String moduleName, List<CourseModuleResourceMapping> mappings, List<Resource> resources) {
+    public Module(int moduleId, Course course, String moduleName, List<CourseModuleResourceMapping> mappings, List<Resource> resources) {
         this.moduleId = moduleId;
         this.course = course;
         this.moduleName = moduleName;
